Avoid race conditions on data appends. fixes #8
An append to a specific Appendr bin usually translates to a read (old state) +
write (new state) to the external storage service. Therefore, because AppEngine
processes requests concurrently, it is possible that two append requests come
in, both read the same old state, the first request finishes and the other
overwrites the just appended data.

This fix resolves the issue by using task queues. A task is created for each
append request and added to a task queue based on the bin id of the append
request. Since appends to the same bin have the same bin id, all request for the
same bin will go to the same queue. Furthermore, all queues have a bucket size
of 1 which disables concurrent execution of request from the same queue. This
solves the problem of race conditions since requests to the same bin will be
processed serially.

However, in order to improve performance and not limit it to processing a single
request at a time globally, multiple queues are used (currently 10). Therefore,
requests to different bins may be processed concurrently, but requests to
the same bin are processed serially.